quick-fix for azure failing builds
