+++ build.gradle.kts
@@ -16,13 +16,14 @@
 
 plugins {
     base
-    id("com.bmuschko.docker-remote-api")
-    id("org.nosphere.apache.rat")
+    id("com.bmuschko.docker-remote-api") version "6.1.3"
+    id("org.nosphere.apache.rat") version "0.6.0"
 }
 
 repositories {
     jcenter()
     mavenLocal()
+    mavenCentral()
     maven { url = uri("https://plugins.gradle.org/m2/") }
 }
 
@@ -54,5 +55,92 @@ tasks {
 
 }
 
-apply(from = "gradle/docker.gradle.kts")
+
+
+val dockerBaseImageRef = "$buildDir/.docker/buildBaseImage-imageId.txt"
+val dockerBaseAlpineImageRef = "$buildDir/.docker/buildBaseAlpineImage-imageId.txt"
+
+val dockerBaseImageId = "${project.property("docker.domain")}/knotx:${project.version}"
+val dockerBaseAlpineImageId = "${project.property("docker.domain")}/knotx-alpine:${project.version}"
+
+val dockerfileBaseImagePath = "$projectDir/src/main/docker/base/Dockerfile"
+val dockerfileBaseAlpineImagePath = "$projectDir/src/main/docker/base-alpine/Dockerfile"
+
+fun com.bmuschko.gradle.docker.tasks.image.DockerRemoveImage.remove(name: String): Unit {
+    val spec = object : Spec<Task> {
+        override fun isSatisfiedBy(task: Task): Boolean {
+            return file(name).exists()
+        }
+    }
+    onlyIf(spec)
+
+    targetImageId(if (file(name).exists()) file(name).readText() else "")
+    onError {
+        if (!this.message!!.contains("No such image"))
+            throw this
+    }
+}
+
+tasks.register<Copy>("copyDockerfileBase") {
+    group = "docker"
+    from(dockerfileBaseImagePath)
+    into("$buildDir/out/base")
+}
+
+tasks.register<Copy>("copyDockerfileBaseAlpine") {
+    group = "docker"
+    from(dockerfileBaseAlpineImagePath)
+    into("$buildDir/out/baseAlpine")
+}
+
+tasks.register<com.bmuschko.gradle.docker.tasks.image.DockerRemoveImage>("removeBaseImage") {
+    group = "docker"
+    remove(dockerBaseImageRef)
+}
+
+tasks.register<com.bmuschko.gradle.docker.tasks.image.DockerRemoveImage>("removeBaseAlpineImage") {
+    group = "docker"
+    remove(dockerBaseAlpineImageRef)
+}
+
+tasks.register<com.bmuschko.gradle.docker.tasks.image.DockerBuildImage> ("buildBaseImage") {
+    group = "docker"
+    inputDir.set(file("$buildDir/out"))
+    dockerFile.set(file("$buildDir/out/base/Dockerfile"))
+    images.add(dockerBaseImageId)
+    dependsOn("removeBaseImage", "copyDockerfileBase")
+}
+
+tasks.register<com.bmuschko.gradle.docker.tasks.image.DockerBuildImage> ("buildBaseAlpineImage") {
+    group = "docker"
+    inputDir.set(file("$buildDir/out"))
+    dockerFile.set(file("$buildDir/out/baseAlpine/Dockerfile"))
+    images.add(dockerBaseAlpineImageId)
+    dependsOn("removeBaseAlpineImage", "copyDockerfileBaseAlpine")
+}
+
+tasks.register<com.bmuschko.gradle.docker.tasks.image.DockerSaveImage>("saveBaseImage") {
+    destFile.set(file("$projectDir/build/docker/docker-build-base.tar"))
+    image.set(dockerBaseImageId)
+    useCompression.set(true)
+    mustRunAfter("buildBaseImage")
+}
+
+tasks.register<com.bmuschko.gradle.docker.tasks.image.DockerSaveImage>("saveBaseAlpineImage") {
+    destFile.set(file("$projectDir/build/docker/docker-build-base-alpine.tar"))
+    image.set(dockerBaseAlpineImageId)
+    useCompression.set(true)
+    mustRunAfter("buildBaseAlpineImage")
+}
+
+tasks.register("prepareDocker") {
+    dependsOn("buildBaseImage", "buildBaseAlpineImage", "saveDockerImage")
+    mustRunAfter("downloadAndUnzipDistribution")
+}
+
+
+tasks.register("saveDockerImage") {
+    dependsOn("saveBaseImage", "saveBaseAlpineImage")
+}
+
 apply(from = "gradle/distribution.gradle.kts")
